etl/crawler_benefit/pyproject.toml
[tool.poetry]
name = "crawler-benefit"
version = "0.1.0"
description = ""
authors = ["Vinicius Espindola <vinnyaoe@gmail.com>"]
readme = "README.md"
packages = [{include = "crawler_benefit"}]

[tool.poetry.dependencies]
python = "^3.11.4"
beautifulsoup4 = "^4.12.2"
pika = "^1.3.2"
selenium = "^4.11.2"
python-decouple = "^3.8"
redis = "^4.6.0"
elasticsearch = "^8.9.0"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
isort = "^5.12.0"
pytest = "^7.2.1"
pytest-cov = "^4.0.0"
pytest-mock = "^3.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[settings]
known_third_party = ["pytest", "requests"]
[tool.black]
line-length = 120
[tool.isort]
profile = "black"